Rename deploy() method to invoke(), extract changes/mentoring handling interfaces from actual email fetching, move email handling to lib/email. Moreover implement stub code in ImportComments._process_changes

+++ debexpo/lib/filesystem.py
@@ -161,7 +161,7 @@ class CheckFiles(object):
             if os.path.exists(file):
                     log.debug("Removing file '%s'" % (file))
                     os.unlink(file)
-        if os.path.isdir(path):
+        if os.path.isdir(path) and os.listdir(path) == []:
             log.debug("Remove empty package repository '%s'" % (path))
             os.rmdir(path)
